Loading screen for calibration added
Minor cleaning of the code, next will be documentation. The main is one big pile, I will do some minor cleanups, but I don't find it usefull to make a seperate class for snake since it is just for the sake of this demo to show that I can use the gyro as a controller and have snake displaying the result on the oled display. It isn't by any means code that can be used by someone else. The library is, but has some snake specific function I gave it, for the sake of the demo too.